requirements.txt
langgraph>=0.2.60
mcp[cli]>=1.13.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic-ai
fastapi>=0.115.0
uvicorn>=0.30.0
